refactor(spark): pull shared HoodieCatalystExpressionUtils bodies into a common base class
Introduce BaseHoodieCatalystExpressionUtils in hudi-spark-common carrying the
method bodies that are byte-identical across Spark 3.3-4.2 (normalizeExprs,
extractPredicatesWithinOutputSet, canUpCast, and the order-preserving
transformation matcher). The ParseToDate/ParseToTimestamp patterns, whose
case-class shapes differ across Spark versions, are kept per version behind
the unapplyOrderPreservingDateParsing hook.

HoodieSpark4CatalystExpressionUtils in hudi-spark4-common now carries the
bodies uniform within the Spark 4.x family (getEncoder, matchCast,
unapplyCastExpression, date-parsing hook), turning the Spark 4.x objects into
empty declarations. The Spark 3.x objects keep only genuine per-version
deltas (RowEncoder vs ExpressionEncoder, AnsiCast, EvalMode). The empty
HoodieSpark3CatalystExpressionUtils shim is removed; adapter-facing object
names are unchanged.
